Refactoring Website Next components WIP
