+++ site/components/event-image/style.css
