+++ site/components/icons/style.css
