+++ site/routes/index.js
@@ -11,9 +11,9 @@ import OfflinePage from '../pages/offline';
 import AboutUsPage from '../pages/about-us';
 
 import JoinUsPage from '../pages/join-us';
-import JobPage from '../../website-next/src/shared/containers/job';
-import Events from '../../website-next/src/shared/containers/events';
-import EventPage from '../../website-next/src/shared/containers/event';
+import JobPage from '../pages/job';
+import Events from '../pages/events';
+import EventPage from '../pages/event';
 
 const componentMap = {
   homePage: HomePage,
